Metrics return longs

+++ src/java/fault/circuit/DefaultCircuitBreaker.java
@@ -64,9 +64,9 @@ public class DefaultCircuitBreaker implements CircuitBreaker {
             if (state.get() == CLOSED) {
                 BreakerConfig config = this.breakerConfig.get();
                 int timePeriod = config.timePeriodInMillis / 1000;
-                int errorsForTimePeriod = actionMetrics.getMetricCountForTimePeriod(Metric.ERROR, timePeriod);
-                int timeoutsForTimePeriod = actionMetrics.getMetricCountForTimePeriod(Metric.TIMEOUT, timePeriod);
-                int failuresForTimePeriod = errorsForTimePeriod + timeoutsForTimePeriod;
+                long errorsForTimePeriod = actionMetrics.getMetricCountForTimePeriod(Metric.ERROR, timePeriod);
+                long timeoutsForTimePeriod = actionMetrics.getMetricCountForTimePeriod(Metric.TIMEOUT, timePeriod);
+                long failuresForTimePeriod = errorsForTimePeriod + timeoutsForTimePeriod;
                 if (config.failureThreshold < failuresForTimePeriod) {
                     lastTestedTime.set(systemTime.currentTimeMillis());
                     state.compareAndSet(CLOSED, OPEN);
